+++ app/src/main/java/org/d3if3118/minikasir/model/Purchase.kt
@@ -3,8 +3,8 @@ package org.d3if3118.minikasir.model
 data class Purchase(
     val namaPelanggan: String,
     val namaBarang: String,
-    val jumlahBeli: Int,
-    val harga: Int,
-    val uangBayar: Int
+    val jumlahBeli: Float,
+    val harga: Float,
+    val uangBayar: Float
 )
