+++ app/src/main/java/org/d3if3118/minikasir/model/PurchaseResult.kt
@@ -1,12 +1,13 @@
 package org.d3if3118.minikasir.model
 
 data class PurchaseResult(
-    val totalBelanja: Int,
-    val uangKembali: Int,
+    val totalBelanja: Float,
+    val uangKembali: Float,
     val bonusType: BonusType,
     val keterangan: String
 ) {
     fun getTotalBelanjaText(): String {
         return "Total belanja: Rp. $totalBelanja"
+
     }
 }
